+++ Demo/Application/Base/AppDelegate.swift
@@ -4,7 +4,6 @@ import BraintreeCore
 @UIApplicationMain class AppDelegate: UIResponder, UIApplicationDelegate {
         
     private let returnURLScheme = "com.braintreepayments.Demo.payments"
-    private let universalLinkURL = "https://braintree-ios-demo.fly.dev/braintree-payments"
     private let processInfoArgs = ProcessInfo.processInfo.arguments
     private let userDefaults = UserDefaults.standard
     
@@ -12,7 +11,6 @@ import BraintreeCore
         registerDefaultsFromSettings()
         persistDemoSettings()
         BTAppContextSwitcher.sharedInstance.returnURLScheme = returnURLScheme
-        BTAppContextSwitcher.sharedInstance.universalLink = universalLinkURL
 
         userDefaults.setValue(true, forKey: "magnes.debug.mode")